refactor: CSVLoader moved to dataframe/loaders.py as DataFrameReadCSV
