+++ src/workflowlib/__init__.py
@@ -1,5 +1,5 @@
 # set package version
 __version__ = '0.1.4'
 
-from . import loaders, serializers, transforms
+from . import loaders, serializers, dataframes
 from .workflow import run, Workflow
