+++ src/workflowlib/dataframes/selection.py
